requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
paho-mqtt
python-dotenv
